+++ tools/oeutil/host/get_endorsements.h
@@ -1,0 +1,9 @@
+// Copyright (c) Open Enclave SDK contributors.
+// Licensed under the MIT License.
+
+#ifndef _OE_UTIL_GET_ENDORSEMENTS_H
+#define _OE_UTIL_GET_ENDORSEMENTS_H
+
+int oeutil_get_endorsements(int argc, const char* argv[]);
+
+#endif // _OE_UTIL_GET_ENDORSEMENTS_H